refactor(types): migrate from tsd to tstyche
